requirements.txt
Flask==3.0.3
torch==2.8.0
torchvision==0.23.0
numpy==2.1.1
Pillow==11.3.0
webcolors==24.11.1
colorthief==0.2.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
